+++ app/src/main/java/com/example/moves/data/database/PosterDb.kt
@@ -1,5 +1,3 @@
 package com.example.moves.data.database
 
-import java.io.Serializable
-
-class PosterDb(val url: String) : Serializable
+class PosterDb(val url: String)
