+++ app/src/main/java/com/example/moves/data/database/RatingDb.kt
@@ -1,6 +1,5 @@
 package com.example.moves.data.database
 
 import com.google.gson.annotations.SerializedName
-import java.io.Serializable
 
-class RatingDb(@SerializedName("kp") val kp: Double) : Serializable
+class RatingDb(@SerializedName("kp") val kp: Double)
