+++ app/src/main/java/com/example/moves/data/network/model/MovieDto.kt
@@ -1,8 +1,6 @@
 package com.example.moves.data.network.model
 
-import androidx.room.Embedded
 import com.google.gson.annotations.SerializedName
-import java.io.Serializable
 
 class MovieDto (
 
@@ -19,10 +17,8 @@ class MovieDto (
     val year: Int,
 
     @SerializedName("poster")
-    @Embedded
     val poster: PosterDto,
 
     @SerializedName("rating")
-    @Embedded
     val rating: RatingDto
-) : Serializable
+)
