+++ app/src/main/java/com/example/moves/data/network/model/PosterDto.kt
@@ -1,5 +1,3 @@
 package com.example.moves.data.network.model
 
-import java.io.Serializable
-
-class PosterDto(val url: String) : Serializable
+class PosterDto(val url: String)
