+++ app/src/main/java/com/example/moves/data/network/model/RatingDto.kt
@@ -1,6 +1,5 @@
 package com.example.moves.data.network.model
 
 import com.google.gson.annotations.SerializedName
-import java.io.Serializable
 
-class RatingDto(@SerializedName("kp") val kp: Double) : Serializable
+class RatingDto(@SerializedName("kp") val kp: Double)
