--- app/src/main/java/com/example/moves/domain/Movie.kt
@@ -1,30 +1,0 @@
-package com.example.moves.domain
-
-import androidx.room.Embedded
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-class Movie(
-
-    @SerializedName("id")
-    val id: Int,
-
-    @SerializedName("name")
-    val name: String,
-
-    @SerializedName("description")
-    val description: String,
-
-    @SerializedName("year")
-    val year: Int,
-
-    @SerializedName("poster")
-    @Embedded
-    val poster: Poster,
-
-    @SerializedName("rating")
-    @Embedded
-    val rating: Rating
-) : Serializable
