--- app/src/main/java/com/example/moves/domain/MovieResponse.kt
@@ -1,5 +1,0 @@
-package com.example.moves.domain
-
-import com.google.gson.annotations.SerializedName
-
-class MovieResponse(@SerializedName("docs") val movies: List<Movie>)
