--- app/src/main/java/com/example/moves/domain/Rating.kt
@@ -1,6 +1,0 @@
-package com.example.moves.domain
-
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-class Rating(@SerializedName("kp") val kp: Double) : Serializable
