--- app/src/main/java/com/example/moves/domain/ReviewResponse.kt
@@ -1,5 +1,0 @@
-package com.example.moves.domain
-
-import com.google.gson.annotations.SerializedName
-
-class ReviewResponse(@SerializedName("docs") val reviewList: List<Review>)
