--- app/src/main/java/com/example/moves/domain/Trailer.kt
@@ -1,16 +1,0 @@
-package com.example.moves.domain
-
-import com.google.gson.annotations.SerializedName
-
-class Trailer(
-
-    @SerializedName("name")
-    val name: String,
-
-    @SerializedName("url")
-    val url: String
-) {
-    override fun toString(): String {
-        return "movie {name=$name, url=$url}"
-    }
-}
