--- app/src/main/java/com/example/moves/domain/TrailerResponse.kt
@@ -1,5 +1,0 @@
-package com.example.moves.domain
-
-import com.google.gson.annotations.SerializedName
-
-class TrailerResponse(@SerializedName("videos") val trailsList: Video)
