--- app/src/main/java/com/example/moves/domain/Video.kt
@@ -1,5 +1,0 @@
-package com.example.moves.domain
-
-import com.google.gson.annotations.SerializedName
-
-class Video(@SerializedName("trailers") val trailers: List<Trailer>)
