+++ app/src/main/java/com/example/moves/domain/model/MovieResponse.kt
@@ -1,0 +1,3 @@
+package com.example.moves.domain.model
+
+class MovieResponse(val movies: List<Movie>)
