+++ app/src/main/java/com/example/moves/domain/model/Poster.kt
@@ -1,4 +1,4 @@
-package com.example.moves.domain
+package com.example.moves.domain.model
 
 import java.io.Serializable
