+++ app/src/main/java/com/example/moves/domain/model/Rating.kt
@@ -1,0 +1,5 @@
+package com.example.moves.domain.model
+
+import java.io.Serializable
+
+class Rating(val kp: Double) : Serializable
