+++ app/src/main/java/com/example/moves/domain/model/Review.kt
@@ -1,0 +1,7 @@
+package com.example.moves.domain.model
+
+class Review(
+    val title: String,
+    val type: String,
+    val review: String,
+)
