+++ app/src/main/java/com/example/moves/domain/model/ReviewResponse.kt
@@ -1,0 +1,3 @@
+package com.example.moves.domain.model
+
+class ReviewResponse(val reviewList: List<Review>)
