+++ app/src/main/java/com/example/moves/domain/model/Trailer.kt
@@ -1,0 +1,6 @@
+package com.example.moves.domain.model
+
+class Trailer(
+    val name: String,
+    val url: String
+)
