+++ app/src/main/java/com/example/moves/domain/model/TrailerResponse.kt
@@ -1,0 +1,3 @@
+package com.example.moves.domain.model
+
+class TrailerResponse(val trailsList: Video)
