+++ app/src/main/java/com/example/moves/domain/model/Video.kt
@@ -1,0 +1,3 @@
+package com.example.moves.domain.model
+
+class Video(val trailers: List<Trailer>)
