+++ app/src/main/java/com/example/moves/domain/repository/MovieRepository.kt
@@ -1,7 +1,10 @@
-package com.example.moves.domain
+package com.example.moves.domain.repository
 
 import androidx.lifecycle.LiveData
-import com.example.moves.data.network.model.*
+import com.example.moves.domain.model.Movie
+import com.example.moves.domain.model.MovieResponse
+import com.example.moves.domain.model.ReviewResponse
+import com.example.moves.domain.model.TrailerResponse
 import io.reactivex.rxjava3.core.Completable
 import io.reactivex.rxjava3.core.Single
