+++ app/src/main/java/com/example/moves/domain/usecase/AddMovieInDatabaseUseCase.kt
@@ -1,4 +1,7 @@
-package com.example.moves.domain
+package com.example.moves.domain.usecase
+
+import com.example.moves.domain.model.Movie
+import com.example.moves.domain.repository.MovieRepository
 
 class AddMovieInDatabaseUseCase(private val repository: MovieRepository) {
