+++ app/src/main/java/com/example/moves/domain/usecase/LoadMovieTrailersUseCase.kt
@@ -1,4 +1,6 @@
-package com.example.moves.domain
+package com.example.moves.domain.usecase
+
+import com.example.moves.domain.repository.MovieRepository
 
 class LoadMovieTrailersUseCase (private val repository: MovieRepository) {
