+++ app/src/main/java/com/example/moves/presentation/adapters/MovieRAdapter.kt
@@ -8,12 +8,12 @@ import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.moves.R
-import com.example.moves.domain.Movie
+import com.example.moves.domain.model.Movie
 import kotlinx.android.synthetic.main.movie_item.view.*
 
 class MovieRAdapter : RecyclerView.Adapter<MovieRAdapter.MovieViewHolder>() {
 
-    var movies: List<Movie> = ArrayList<Movie>()
+    var movies: List<Movie> = ArrayList()
 
     private lateinit var onReachEndListener: OnReachEndListener
