+++ app/src/main/java/com/example/moves/presentation/screens/detail/DetailMoveActivity.kt
@@ -11,8 +11,8 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.moves.R
-import com.example.moves.domain.Movie
-import com.example.moves.domain.Trailer
+import com.example.moves.domain.model.Movie
+import com.example.moves.domain.model.Trailer
 import com.example.moves.presentation.adapters.ReviewRAdapter
 import com.example.moves.presentation.adapters.TrailerRAdapter
 import kotlinx.android.synthetic.main.activity_detail_move.*
