+++ app/src/main/java/com/example/moves/presentation/screens/favorite/FavoriteMoviesViewModel.kt
@@ -3,11 +3,9 @@ package com.example.moves.presentation.screens.favorite
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
-import com.example.moves.data.database.MovieDatabase
-import com.example.moves.data.database.MoviesDao
 import com.example.moves.data.repository.MovieRepositoryImpl
-import com.example.moves.domain.GetFavoriteMovesListUseCase
-import com.example.moves.domain.Movie
+import com.example.moves.domain.usecase.GetFavoriteMovesListUseCase
+import com.example.moves.domain.model.Movie
 
 class FavoriteMoviesViewModel(application: Application) : AndroidViewModel(application) {
